Added data route handler to fetch post data for a user, modified get-posts route handler to only fetch encoded buffer (as base64 string) for one file at a time from S3, renamed route handler to get-post-files.
