+++ posts/src/routes/get-post-files.ts
@@ -1,0 +1,54 @@
+import express, { Request, Response } from 'express';
+import {
+  currentUser,
+  requireAuth,
+  BadRequestError,
+} from '@ggabella-photo-share/common';
+import { AWS } from '../index';
+import { S3 } from 'aws-sdk';
+import { encode } from '../utils/encode';
+
+const router = express.Router();
+
+router.get(
+  '/api/postFiles',
+  requireAuth,
+  currentUser,
+  async (req: Request, res: Response) => {
+    const user = req.currentUser;
+
+    const { postKey } = req.body;
+
+    if (!user) {
+      throw new BadRequestError(
+        'No user signed in. Please sign in to access this route.'
+      );
+    }
+
+    const s3 = new AWS.S3();
+
+    const fetchParams: S3.Types.GetObjectRequest = {
+      Bucket: 'photo-share-app',
+      Key: postKey,
+    };
+
+    s3.getObject(fetchParams, async (err, data) => {
+      if (err) {
+        console.log(err);
+        throw new Error('Error fetching the photo');
+      }
+      if (data) {
+        console.log(data);
+        const buffer = data.Body;
+
+        const convertedFileString = encode(buffer as Buffer);
+
+        // const image = `<img src='data:image/jpeg;base64,${convertedFileString}'/>`;
+
+        res.send(convertedFileString);
+      }
+    });
+  }
+);
+
+export { router as getPostFilesRouter };
